// src/app/shared/enums/ExerciseEnum.ts
export enum ExerciseEnum {
  SNATCH = 'SNATCH',
  POWER_SNATCH = 'POWER_SNATCH',
  HANG_SNATCH = 'HANG_SNATCH',
  HANG_POWER_SNATCH = 'HANG_POWER_SNATCH',
  CLEAN = 'CLEAN',
  POWER_CLEAN = 'POWER_CLEAN',
  HANG_SQUAT_CLEAN = 'HANG_SQUAT_CLEAN',
  HANG_POWER_CLEAN = 'HANG_POWER_CLEAN',
  CLEAN_AND_JERK = 'CLEAN_AND_JERK',
  SPLIT_JERK = 'SPLIT_JERK',
  STRICT_PRESS = 'STRICT_PRESS',
  PUSH_PRESS = 'PUSH_PRESS',
  PUSH_JERK = 'PUSH_JERK',
  BENCH_PRESS = 'BENCH_PRESS',
  FRONT_SQUAT = 'FRONT_SQUAT',
  BACK_SQUAT = 'BACK_SQUAT',
  OVERHEAD_SQUAT = 'OVERHEAD_SQUAT',
  DEADLIFT = 'DEADLIFT',
  THRUSTER = 'THRUSTER',
  NONE = 'NONE',
}